feat: permit area picture extension
* minor: format code

* poja: regen with new env vars

* test: add tile.extender.baseurl env var to Test EnvConf

* wip: permit area picture extension

* test: data mismatch -> guessedLayer was from charente department but mockdata was from ain department

* refactor: WmsImageSource.downloadImage has only one param (AreaPicture) and WmsImageSource::supports checks whether image source supports given source or not

* wip: TileExtenderImageSource

* chore: FileDownloader supports POST and GET methods

refactor: FileDownloader Impl to @Component

* wip: tile extender

* minor: format code

+++ src/main/java/app/bpartners/api/service/WMS/imageSource/WmsImageSourceFacade.java
@@ -1,23 +1,39 @@
 package app.bpartners.api.service.WMS.imageSource;
 
+import app.bpartners.api.file.FileDownloader;
+import app.bpartners.api.model.AreaPicture;
 import app.bpartners.api.model.AreaPictureMapLayer;
 import app.bpartners.api.service.WMS.AreaPictureMapLayerService;
 import app.bpartners.api.service.WMS.Tile;
 import app.bpartners.api.service.WMS.imageSource.exception.BlankImageException;
 import java.io.File;
 import java.net.URI;
-import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
 @Component
-@AllArgsConstructor
 @Primary
 final class WmsImageSourceFacade extends AbstractWmsImageSource {
   private final OpenStreetMapImageSource openStreetMapImageSource;
   private final GeoserverImageSource geoserverImageSource;
   private final AreaPictureMapLayerService areaPictureMapLayerService;
+  private final TileExtenderImageSource tileExtenderImageSource;
   private final ImageValidator imageValidator;
+
+  private WmsImageSourceFacade(
+      FileDownloader fileDownloader,
+      OpenStreetMapImageSource openStreetMapImageSource,
+      GeoserverImageSource geoserverImageSource,
+      AreaPictureMapLayerService areaPictureMapLayerService,
+      TileExtenderImageSource tileExtenderImageSource,
+      ImageValidator imageValidator) {
+    super(fileDownloader);
+    this.openStreetMapImageSource = openStreetMapImageSource;
+    this.geoserverImageSource = geoserverImageSource;
+    this.areaPictureMapLayerService = areaPictureMapLayerService;
+    this.tileExtenderImageSource = tileExtenderImageSource;
+    this.imageValidator = imageValidator;
+  }
 
   @Override
   protected URI getURI(Tile tile, AreaPictureMapLayer areaPictureMapLayer) {
@@ -28,21 +44,28 @@ final class WmsImageSourceFacade extends AbstractWmsImageSource {
   }
 
   @Override
-  public File downloadImage(String filename, Tile tile, AreaPictureMapLayer areaPictureMapLayer) {
-    return switch (areaPictureMapLayer.getSource()) {
-      case OPENSTREETMAP -> openStreetMapImageSource.downloadImage(
-          filename, tile, areaPictureMapLayer);
+  public File downloadImage(AreaPicture areaPicture) {
+    if (areaPicture.isExtended()) {
+      return tileExtenderImageSource.downloadImage(areaPicture);
+    }
+    return switch (areaPicture.getCurrentLayer().getSource()) {
+      case OPENSTREETMAP -> openStreetMapImageSource.downloadImage(areaPicture);
       case GEOSERVER -> {
-        var file = geoserverImageSource.downloadImage(filename, tile, areaPictureMapLayer);
+        var file = geoserverImageSource.downloadImage(areaPicture);
         try {
           imageValidator.accept(file);
           yield file;
         } catch (BlankImageException e) {
           file.delete();
-          yield openStreetMapImageSource.downloadImage(
-              filename, tile, areaPictureMapLayerService.getDefaultLayer());
+          areaPicture.setCurrentLayer(areaPictureMapLayerService.getDefaultLayer());
+          yield openStreetMapImageSource.downloadImage(areaPicture);
         }
       }
     };
+  }
+
+  @Override
+  public boolean supports(AreaPicture areaPicture) {
+    return true;
   }
 }
